feature: implement check-in rate tracking and stats

+++ backend/src/main/java/com/example/backend/repository/ShowingTimeRepository.java
@@ -1,6 +1,5 @@
 package com.example.backend.repository;
 
-import com.example.backend.model.Address;
 import com.example.backend.model.ShowingTime;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -11,5 +10,6 @@ import java.util.Optional;
 @Repository
 public interface ShowingTimeRepository extends JpaRepository<ShowingTime,Integer> {
     @EntityGraph(attributePaths = {"seats", "zones"})
-    Optional<ShowingTime> findWithLayoutById(Long id);
+    Optional<ShowingTime> findWithLayoutById(Integer id);
+
 }
